Moved RequestPermissionsAPI to a separate class
